# services/gateway/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SessionResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "EMAIL_TAKEN",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "EMAIL_TAKEN"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SessionResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "INVALID_CREDENTIALS"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_LOCKED\n\nACCOUNT_DEACTIVATED",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "ACCOUNT_DEACTIVATED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/firebase": {
      "post": {
        "operationId": "AuthController_firebase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseSignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SessionResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "FIREBASE_TOKEN_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "FIREBASE_TOKEN_INVALID"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_LOCKED\n\nACCOUNT_DEACTIVATED",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "ACCOUNT_DEACTIVATED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RefreshResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "UNAUTHENTICATED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "BrewLite API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "fullName": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": ["CUSTOMER", "STAFF", "ADMIN"]
              },
              "permissions": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "menu.manage",
                    "stock.update",
                    "order.board.read",
                    "order.status.update",
                    "order.cancel.paid",
                    "order.read.all",
                    "promotion.manage",
                    "user.manage",
                    "report.read"
                  ]
                }
              },
              "hasPassword": {
                "type": "boolean"
              },
              "preferredLocale": {
                "type": "string",
                "enum": ["en", "vi"]
              },
              "createdAt": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "fullName",
              "role",
              "permissions",
              "hasPassword",
              "preferredLocale",
              "createdAt"
            ]
          },
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "preferredLocale": {
            "type": "string",
            "enum": ["en", "vi"]
          }
        },
        "required": ["email", "password", "fullName"],
        "additionalProperties": false
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          }
        },
        "required": ["email", "password"],
        "additionalProperties": false
      },
      "FirebaseSignInDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          }
        },
        "required": ["idToken"],
        "additionalProperties": false
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string"
          }
        },
        "required": ["accessToken", "accessTokenExpiresAt"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          }
        },
        "required": ["refreshToken"],
        "additionalProperties": false
      },
...
    }
  }
}